Safetensors CLI check (#776)
* Add safetensors availability check for CLI users

* initial draft revision

* draft revision

* Update tests/test_benchmark_modules/test_hf.py

Comment from Dan:

This test seems to give false confidence, since currently the BenchmarkConfig does not have an only_allow_safetensors attribute.

There is already a benchmark_config fixture, so you could just use that as follows:

Co-authored-by: Dan Saattrup Nielsen <47701536+saattrupdan@users.noreply.github.com>

* Update src/scandeval/benchmark_modules/hf.py

Comment from Dan:

No need to log separately, as this is done automatically when InvalidModel is raised. Instead, we can just add the extra information into the exception body, and also condition it on whether we're using the CLI or not:

Co-authored-by: Dan Saattrup Nielsen <47701536+saattrupdan@users.noreply.github.com>

* Update src/scandeval/data_models.py

Having a default value here is error-prone - we do that elsewhere to avoid silent errors (which is currently what is happening now).

Co-authored-by: Dan Saattrup Nielsen <47701536+saattrupdan@users.noreply.github.com>

* adding missing attributes and docstrings

* added missing new arguments

---------

Co-authored-by: Dan Saattrup Nielsen <47701536+saattrupdan@users.noreply.github.com>